Feature #2 - Use case to findVendor with ID "vendorId" to verify the Multi tenancy FindById implementation in MultiTenancyRepository + integration tests to verify the correctness.
