+++ src/main/java/com/github/mdevloo/multi/tenancy/core/inventory/usecase/inventory/CreateNewInventory.java
@@ -1,7 +1,7 @@
-package com.github.mdevloo.multi.tenancy.core.inventory.usecase;
+package com.github.mdevloo.multi.tenancy.core.inventory.usecase.inventory;
 
-import com.github.mdevloo.multi.tenancy.core.inventory.domain.Inventory;
-import com.github.mdevloo.multi.tenancy.core.inventory.domain.InventoryRepository;
+import com.github.mdevloo.multi.tenancy.core.inventory.domain.inventory.Inventory;
+import com.github.mdevloo.multi.tenancy.core.inventory.domain.inventory.InventoryRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
