+++ src/main/java/com/github/mdevloo/multi/tenancy/core/inventory/usecase/inventory/FindInventory.java
@@ -1,10 +1,11 @@
-package com.github.mdevloo.multi.tenancy.core.inventory.usecase;
+package com.github.mdevloo.multi.tenancy.core.inventory.usecase.inventory;
 
-import com.github.mdevloo.multi.tenancy.core.inventory.domain.Inventory;
-import com.github.mdevloo.multi.tenancy.core.inventory.domain.InventoryRepository;
+import com.github.mdevloo.multi.tenancy.core.inventory.domain.inventory.Inventory;
+import com.github.mdevloo.multi.tenancy.core.inventory.domain.inventory.InventoryRepository;
 import com.github.mdevloo.multi.tenancy.fwk.ObjectNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.UUID;
 
@@ -14,6 +15,7 @@ public class FindInventory {
 
   private final InventoryRepository inventoryRepository;
 
+  @Transactional
   public Inventory execute(final UUID id) {
     return this.inventoryRepository
         .findById(id)
