+++ src/test/java/com/github/mdevloo/multi/tenancy/core/inventory/usecase/inventory/FindInventoryTest.java
@@ -1,7 +1,7 @@
-package com.github.mdevloo.multi.tenancy.core.inventory.usecase;
+package com.github.mdevloo.multi.tenancy.core.inventory.usecase.inventory;
 
 import com.github.mdevloo.multi.tenancy.AbstractIntegrationTest;
-import com.github.mdevloo.multi.tenancy.core.inventory.domain.Inventory;
+import com.github.mdevloo.multi.tenancy.core.inventory.domain.inventory.Inventory;
 import com.github.mdevloo.multi.tenancy.fwk.ObjectNotFoundException;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
